delete

+++ src/app/layout/client/listing-client/listing-client.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ClientService } from '../../../core/services/client.services';
 import { Router } from '@angular/router';
 import { ToasterService } from 'angular2-toaster';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-listing-client',
@@ -14,9 +15,9 @@ export class ListingClientComponent implements OnInit {
   config: any;
 
   constructor(private router: Router,
-    private toast: ToasterService,private clientService: ClientService) { }
+    private toast: ToasterService, private clientService: ClientService) { }
 
-  ngOnInit(){
+  ngOnInit() {
     this.getClient();
     this.config = {
       itemsPerPage: 10,
@@ -27,29 +28,45 @@ export class ListingClientComponent implements OnInit {
   getClient() {
     this.clientService.getClient().subscribe((response) => {
       this.clientDetail = response.results;
-   });
-  } 
-   
-  pageChanged(event:any){
-    this.config.currentPage = event;
-  }
-  updateclient(client: any){
-    this.router.navigate(['/client/update', client.id]);
+    });
   }
 
-  deleteclient(client: any) {
-    this.clientService.deleteClient(client.id)
-      .subscribe(
-        (response: any) => {
-          const index = this.clientDetail.indexOf(client, 0);
-          if (index > -1) {
-            this.clientDetail.splice(index, 1);
-            this.getClient();
-            this.toast.pop('success', 'Your client delete is not recover');
-          }
-          else {
-            this.toast.pop('error', 'Your client is safe!');
-          }
-        });
+  pageChanged(event: any) {
+    this.config.currentPage = event;
   }
+  updateclient(client: any) {
+    this.router.navigate(['/client/update', client.id]);
+  }
+  
+  deleteclient(client: any) {
+    Swal.fire({
+      title: '<i class="fa fa-trash" aria-hidden="true"></i>',
+      text: 'Are you sure you want to delete?',
+      showCancelButton: true,
+      icon: 'warning',
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    })
+      .then((result) => {
+        if (result.value) {
+          this.clientService.deleteClient(client.id)
+            .subscribe(
+              (response: any) => {
+                const index = this.clientDetail.indexOf(client, 0);
+                if (index > -1) {
+                  this.clientDetail.splice(index, 1);
+                  this.getClient();
+                  Swal.fire(
+                    'Client is Deleted!',
+                    'success'
+                  )
+                }
+              });
+        } else {
+          Swal.fire('Your Client is safe!');
+        }
+      });
+  }
+
 }
