+++ src/app/layout/expense/listing-expense/listing-expense.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ExpenseService } from '../../../core/services/expense.services';
 import { ToasterService } from 'angular2-toaster';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-listing-expense',
@@ -39,18 +40,33 @@ export class ListingExpenseComponent implements OnInit {
     }
 
     deleteExpense(expense: any) {
-      this.expenseService.deleteExpense(expense.id)
-        .subscribe(
-          (response: any) => {
-            const index = this.expense.indexOf(expense, 0);
-            if (index > -1) {
-              this.expense.splice(index, 1);
-              this.getExpense();
-              this.toast.pop('success', 'Your Expense delete is not recover');
-            }
-            else {
-              this.toast.pop('error', 'Your Expense is safe!');
-            }
-          });
+      Swal.fire({
+        title: '<i class="fa fa-trash" aria-hidden="true"></i>',
+        text: 'Are you sure you want to delete?',
+        showCancelButton: true,
+        icon: 'warning',
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      })
+        .then((result) => {
+          if (result.value) {
+              this.expenseService.deleteExpense(expense.id)
+              .subscribe(
+                (response: any) => {
+                  const index = this.expense.indexOf(expense, 0);
+                  if (index > -1) {
+                    this.expense.splice(index, 1);
+                    this.getExpense();
+                    Swal.fire(
+                      'Expense is Deleted!',
+                      'success'
+                    )
+                  }
+                });
+          } else {
+            Swal.fire('Your Expense is safe!');
+          }
+        });
     }
 }
