+++ src/app/layout/products/deliverable-product/listing-end-product/listing-end-product.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToasterService } from 'angular2-toaster';
 import { EndProductService } from '../../../../core/services/end-product.services';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-listing-end-product',
@@ -38,19 +39,35 @@ export class ListingEndProductComponent implements OnInit {
   pageChanged(event:any){
     this.config.currentPage = event;
   }
+
   deleteProduct(product: any) {
-    this.endProductService.deleteEndProducts(product.id)
-      .subscribe(
-        (response: any) => {
-          const index = this.endProduct.indexOf(product, 0);
-          if (index > -1) {
-            this.endProduct.splice(index, 1);
-            this.getEndProduct();
-            this.toast.pop('success', 'Your Deliverable Product delete is not recover');
-          }
-          else {
-            this.toast.pop('error', 'Your Deliverable Product is safe!');
-          }
-        });
+    Swal.fire({
+      title: '<i class="fa fa-trash" aria-hidden="true"></i>',
+      text: 'Are you sure you want to delete?',
+      showCancelButton: true,
+      icon: 'warning',
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    })
+      .then((result) => {
+        if (result.value) {
+            this.endProductService.deleteEndProducts(product.id)
+            .subscribe(
+              (response: any) => {
+                const index = this.endProduct.indexOf(product, 0);
+                if (index > -1) {
+                  this.endProduct.splice(index, 1);
+                  this.getEndProduct();
+                  Swal.fire(
+                    'Deliverable Product is Deleted!',
+                    'success'
+                  )
+                }
+              });
+        } else {
+          Swal.fire('Your Deliverable Product is safe!');
+        }
+      });
   }
 }
