+++ src/app/layout/products/raw-material/listing-raw-materials/listing-raw-materials.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { RawMaterialService } from '../../../../core/services/materials.services';
 import { Router } from '@angular/router';
 import { ToasterService } from 'angular2-toaster';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-listing-raw-materials',
@@ -37,20 +38,36 @@ export class ListingRawMaterialsComponent implements OnInit {
   pageChanged(event:any){
     this.config.currentPage = event;
   }
-  deleteMaterial(material: any) {
-    this.materialService.deleteMaterial(material.id)
-      .subscribe(
-        (response: any) => {
-          const index = this.rawMaterial.indexOf(material, 0);
-          if (index > -1) {
-            this.rawMaterial.splice(index, 1);
-            this.getMaterial();
-            this.toast.pop('success', 'Your Material delete is not recover');
-          }
-          else {
-            this.toast.pop('error', 'Your Material is safe!');
-          }
-        });
-  }
 
+
+  deleteMaterial(material: any) {
+    Swal.fire({
+      title: '<i class="fa fa-trash" aria-hidden="true"></i>',
+      text: 'Are you sure you want to delete?',
+      showCancelButton: true,
+      icon: 'warning',
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    })
+      .then((result) => {
+        if (result.value) {
+            this.materialService.deleteMaterial(material.id)
+            .subscribe(
+              (response: any) => {
+                const index = this.rawMaterial.indexOf(material, 0);
+                if (index > -1) {
+                  this.rawMaterial.splice(index, 1);
+                  this.getMaterial();
+                  Swal.fire(
+                    'Raw Material is Deleted!',
+                    'success'
+                  )
+                }
+              });
+        } else {
+          Swal.fire('Your Raw Material is safe!');
+        }
+      });
+  }
 }
