+++ src/app/layout/purchase/listing-purchase/listing-purchase.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ToasterService } from 'angular2-toaster';
 import { Router } from '@angular/router';
 import { PurchaseService } from '../../../core/services/purchase.services';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-listing-purchase',
@@ -39,18 +40,34 @@ export class ListingPurchaseComponent implements OnInit {
     }
 
     deletePurchase(purchase: any) {
-      this.purchaseService.deletePurchase(purchase.id)
-        .subscribe(
-          (response: any) => {
-            const index = this.purchase.indexOf(purchase, 0);
-            if (index > -1) {
-              this.purchase.splice(index, 1);
-              this.getPurchase();
-              this.toast.pop('success', 'Your Purchase delete is not recover');
-            }
-            else {
-              this.toast.pop('error', 'Your Purchase is safe!');
-            }
-          });
+      Swal.fire({
+        title: '<i class="fa fa-trash" aria-hidden="true"></i>',
+        text: 'Are you sure you want to delete?',
+        showCancelButton: true,
+        icon: 'warning',
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      })
+        .then((result) => {
+          if (result.value) {
+              this.purchaseService.deletePurchase(purchase.id)
+              .subscribe(
+                (response: any) => {
+                  const index = this.purchase.indexOf(purchase, 0);
+                  if (index > -1) {
+                    this.purchase.splice(index, 1);
+                    this.getPurchase();
+                    Swal.fire(
+                      'Purchase is Deleted!',
+                      'success'
+                    )
+                  }
+                });
+          } else {
+            Swal.fire('Your Purchase is safe!');
+          }
+        });
     }
+
 }
